Created ProjectList component
